requirements.txt
python-gitlab==3.12.0
GitPython==3.1.30
Jinja2==3.1.2
